feat: split artwork model and add artwork attributes
